refactor(react-gui): take the sequence panel's geometry out of the component
The panel paints residues onto a canvas rather than laying out DOM,
because a long chain is tens of thousands of cells. That makes cell
width, hit testing and the ruler ticks ordinary arithmetic -- but it also
means nothing in the DOM says which residue a click landed on. `pickCell`
is the only answer, and a wrong one selects the wrong residue with
nothing to show for it.

242 lines of that arithmetic, plus the theme read (these draw onto a
canvas, which no stylesheet reaches), move to seqCanvas.ts with no React
in them. The panel keeps the part that is about being a panel: the
selection state, the pointer gestures, the context menu, the layout.

Being callable without a panel is the point, so `pickCell` now has tests:
the row and column a click maps to, the two out-of-range directions, a
click past the last residue of a row, and an insertion-coded residue
(`52A`) still being pickable at column 52 -- the case the leading-integer
parse exists for and the one a rewrite would quietly lose.
